requirements.txt
aenum==3.1.16
alembic==1.16.2
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2025.6.15
cffi==1.17.1
circle-configurations==6.1.0
circle-developer-controlled-wallets==6.1.0
circle-sdk==0.1.0b14
circle-smart-contract-platform==6.1.0
circle-user-controlled-wallets==6.1.0
circle-web3-sdk-util==6.1.0
click==8.2.1
cryptography==45.0.5
dotenv==0.9.9
fastapi==0.98.0
frozendict==2.3.10
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.23.0
pydantic==1.10.22
pydantic_core==2.33.2
python-dateutil==2.7.5
python-dotenv==1.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.54
starlette==0.27.0
typing-inspection==0.4.1
typing_extensions==4.3.0
urllib3==1.26.20
uvicorn==0.35.0
